feat(database): add info to orders tables, new function for adding orders
Split `orders_table` into two tables: `orders_table` for orders, which
contains the time an order was placed, and `order_items` which stores
individiual items associated with an order, containing an ID
referencing the order, an ID referencing the item and the current
status.

The status represents at what stage of being delivered to the customer
the item is, and can be any of 'preparing', 'ready', 'delivered', or
'cancelled'.